Created a central sentence component
