+++ frontend/src/components/Sandbox.js
@@ -1,13 +1,10 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import {
 	HStack,
 	VStack,
 	Box,
-	Center,
 	Icon,
 	Button,
-	Skeleton,
-	Input,
 	Editable,
 	EditablePreview,
 	EditableInput,
@@ -20,6 +17,7 @@ import {
 import { EditIcon, CheckIcon, CloseIcon } from "@chakra-ui/icons";
 import { HiMicrophone } from "react-icons/hi";
 import { BsFillStopFill, BsFillPlayBtnFill } from "react-icons/bs";
+import Sentence from "../components/Sentence";
 import VoiceHistory from "../components/VoiceHistory";
 
 // Global recorder variable stores audio file and current audio state
@@ -130,10 +128,6 @@ const Sandbox = () => {
 	 */
 	const handleStopButtonClick = async () => {
 		audio = await recorder.stop();
-		setAudioUrls([
-			{ sentence: sentence.sentence, url: audio.audioUrl },
-			...audioUrls,
-		]);
 		handleSaveButtonClick();
 		setDisableRecordBtn(false);
 	};
@@ -169,6 +163,7 @@ const Sandbox = () => {
 						setConfidence(r.confidence);
 						setSentenceArr(r.sentence_arr);
 						setTextLoaded(true);
+						setAudioUrls([{url:audio.audioUrl, confidence:r.confidence, sentence_arr:r.sentence_arr}, ...audioUrls]);
 					});
 				} else {
 					console.log(
@@ -179,7 +174,6 @@ const Sandbox = () => {
 		};
 		setDisablePlayBtn(true);
 	};
-	const colours = ["gray", "orange", "red"];
 	return (
 		<HStack
 			height="86vh"
@@ -276,22 +270,7 @@ const Sandbox = () => {
 					borderRadius="3xl"
 					padding="20px"
 				>
-					<Skeleton isLoaded={textLoaded}>
-						<Center
-							color="gray"
-							fontWeight="light"
-							fontSize="3xl"
-							justifyContent="left"
-						>
-							<p>
-								{confidence.map((k, i) => (
-									<span style={{ color: colours[k] }} key={i}>
-										{sentence_arr[i] + " "}
-									</span>
-								))}
-							</p>
-						</Center>
-					</Skeleton>
+					<Sentence confidence={confidence} sentence_arr={sentence_arr} textLoaded={textLoaded} />
 				</Box>
 			</VStack>
 			<VStack height="100%" width="30%">
